feat: split telemetry logging and reading into two files
