+++ internal/utils/telemetry_log.go
@@ -6,7 +6,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/mileusna/useragent"
-	"io"
 	"net"
 	"net/http"
 	"os"
@@ -16,7 +15,7 @@ import (
 
 const logFile = "./data/parser.log"
 
-var mu sync.Mutex
+var telemetryMutex sync.Mutex
 
 type parseTelemetry struct {
 	Success         bool            `json:"success"`
@@ -38,8 +37,8 @@ type userAgent struct {
 // appendLog serializes the provided content to JSON and appends it to the log file
 func appendLog(telemetry parseTelemetry) error {
 	// Lock Mutex
-	mu.Lock()
-	defer mu.Unlock()
+	telemetryMutex.Lock()
+	defer telemetryMutex.Unlock()
 
 	// Create append stream to file
 	file, err := os.OpenFile(logFile, os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0644)
@@ -139,33 +138,4 @@ func LogParseTelemetry(req *http.Request, timings *ProcessTiming, error *APIErro
 		Error:           errorMessage,
 		UserAgent:       ua,
 	})
-}
-
-// ReadTelemetry reads the raw content of the log file and returns it as a byte slice
-func ReadTelemetry() (*[]byte, error) {
-	// Lock Mutex
-	mu.Lock()
-	defer mu.Unlock()
-
-	// Open log file
-	file, err := os.Open(logFile)
-	if err != nil {
-		return nil, err
-	}
-
-	// Close file when done
-	defer func(file *os.File) {
-		err := file.Close()
-		if err != nil {
-			fmt.Println("error while closing log file", err)
-		}
-	}(file)
-
-	// Read file content
-	telemetry, err := io.ReadAll(file)
-	if err != nil {
-		return nil, err
-	}
-
-	return &telemetry, nil
 }
